:abc: :cop:/:zap: v1 docs/info and #15
