+++ docs/modules/.md
@@ -15,13 +15,13 @@ You can also use `$[0]`–`$[n]` for getting nth script argument (compare to bas
 - [configAssign](.md#configassign)
 - [isFIFO](.md#isfifo)
 - [api](.md#api)
-- [read](.md#read)
 - [error](.md#error)
 - [exit](.md#exit)
 - [hasArgs](.md#hasargs)
 
 ### Internal Functions
 
+- [read](.md#read)
 - [Error](.md#error-1)
 
 ### Public Variables
@@ -37,13 +37,10 @@ You can also use `$[0]`–`$[n]` for getting nth script argument (compare to bas
 
 ### Internal Variables
 
-- [nojq](.md#nojq)
-- [nosed](.md#nosed)
-- [noawk](.md#noawk)
+- [version](.md#version)
 
 ### Interfaces
 
-- [ReadOptions](../interfaces/.ReadOptions.md)
 - [STDIN](../interfaces/.STDIN.md)
 
 ## Public Functions
@@ -184,30 +181,6 @@ prog.parse();
 
 ___
 
-### read
-
-▸ **read**(`options`): `Promise`<[`ShellString`](s.md#shellstring)\>
-
-This function mimic [`read`](https://phoenixnap.com/kb/bash-read) command.
-So, the function purpose is reading from `stdin`.
-```js
-const answer= await $.read({ "-p": "Question" });
-const color= await $.read({ "-p": "Your color", completions: [ "red", "green" ] });
-if($.isFIFO(0)) await $.read().then(echo.bind(null, "E.g. for reading received input:"));
-```
-
-#### Parameters
-
-| Name | Type |
-| :------ | :------ |
-| `options` | [`ReadOptions`](../interfaces/.ReadOptions.md) |
-
-#### Returns
-
-`Promise`<[`ShellString`](s.md#shellstring)\>
-
-___
-
 ### error
 
 ▸ **error**(`message`): `Error`
@@ -274,6 +247,28 @@ if($.hasArgs("-v", "--version")) $.exit(0, echo("v0.0.1"));
 ___
 
 ## Internal Functions
+
+### read
+
+▸ **read**(`options`): `Promise`<[`ShellString`](s.md#shellstring)\>
+
+For backward compatibility, **use `s.read()` instead**.
+
+**`Deprecated`**
+
+Use ReadType instead.
+
+#### Parameters
+
+| Name | Type |
+| :------ | :------ |
+| `options` | [`ReadOptions`](../interfaces/s.ReadOptions.md) |
+
+#### Returns
+
+`Promise`<[`ShellString`](s.md#shellstring)\>
+
+___
 
 ### Error
 
@@ -386,30 +381,8 @@ ___
 
 ## Internal Variables
 
-### nojq
+### version
 
-• `Const` **nojq**: ``null`` \| `any`[] \| `Record`<`any`, `any`\>
+• `Const` **version**: `string`
 
-**`Alias`**
-
-[json](../interfaces/.STDIN.md#json)`(null)`
-
-___
-
-### nosed
-
-• `Const` **nosed**: `string`
-
-**`Alias`**
-
-[text](../interfaces/.STDIN.md#text)`("")`
-
-___
-
-### noawk
-
-• `Const` **noawk**: `string`[]
-
-**`Alias`**
-
-[lines](../interfaces/.STDIN.md#lines)`([])`
+Holds current nodejsscript version
